+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - model GroupRole
 - creating group roles for events that need in raid
 - allow to approve in raid with not only main role
-- add "In reserve" status
+- "In reserve" status
+- alternative renderer for event subscribers
 
 ## [0.9.0] - 2019-11-26
 ### Added